refactor(Geom): split InterpolateFace into 3 methods, extract legacy helpers
Split InterpolateFace into three composable stages:
- InterpolateFace(): pure face creation via InterpolateGlobal + populate
  cell2face/face2cell/face2node/faceElemInfo. Delegates to the next two.
- BuildGhostFace(): ghost face determination via evaluateGhostTree(Cell →
  Cell2Face → Face), pull ghost face arrays, convert to local indices.
- MatchFaceBoundary(): match boundaries to faces, communicate cell2face
  and bnd2face ghost data.

InterpolateFace remains the public entry point (calls all three), so
existing callers are unchanged.

Extract legacy face interpolation helpers (EnumerateFacesFromCells,
CollectFaces, CompactFacesAndRemapCell2Face, AssignGhostFacesToCells,
MatchBoundariesToFaces) from Mesh.cpp into Mesh_InterpolateHelpers.hpp.
Move InterpolateFaceLegacy body into Mesh_Legacy.cpp to use them.

Fix uninitialized spanNodes array in Mesh_Elevation.cpp.

20/20 geom tests pass (np=1,2,4). Euler builds clean.
