+++ src/Geom/Mesh_Elevation.cpp
@@ -60,7 +60,7 @@ namespace DNDS::Geom
             for (index iNNode = 0; iNNode < eCell.GetNumElev_O1O2(); iNNode++)
             {
                 auto eSpan = eCell.ObtainElevNodeSpan(iNNode);
-                std::array<index, Elem::CellNumNodeMax> spanNodes;
+                std::array<index, Elem::CellNumNodeMax> spanNodes{UnInitIndex};
                 eCell.ExtractElevNodeSpanNodes(iNNode, c2n, spanNodes);
                 auto spanNodesSrt = spanNodes;
                 std::sort(spanNodesSrt.begin(), spanNodesSrt.begin() + eSpan.GetNumNodes());
